+++ commands/debug/show/generate_origins/generate_origins.go
@@ -8,8 +8,6 @@ import (
 	"context"
 	"fmt"
 	"path"
-	"path/filepath"
-	"strings"
 
 	"github.com/rs/zerolog/log"
 	"github.com/terramate-io/terramate/cloud/api/resources"
@@ -23,10 +21,13 @@ import (
 
 // Spec is the command specification for the generate-origins command.
 type Spec struct {
-	WorkingDir string
-	Engine     *engine.Engine
-	Printers   printer.Printers
-	GitFilter  engine.GitFilter
+	WorkingDir    string
+	Engine        *engine.Engine
+	Printers      printer.Printers
+	GitFilter     engine.GitFilter
+	StatusFilters resources.StatusFilters
+	Tags          []string
+	NoTags        []string
 }
 
 // Name returns the name of the command.
@@ -34,20 +35,19 @@ func (s *Spec) Name() string { return "debug generate-origins" }
 
 // Exec executes the generate-origins command.
 func (s *Spec) Exec(_ context.Context) error {
-	report, err := s.Engine.ListStacks(s.GitFilter, cloudstack.AnyTarget, resources.NoStatusFilters(), false)
+	report, err := s.Engine.ListStacks(s.GitFilter, cloudstack.AnyTarget, s.StatusFilters, false)
 	if err != nil {
 		return errors.E(err, "generate debug: selecting stacks")
 	}
 
-	cfg := s.Engine.Config()
-	selectedStacks := map[project.Path]struct{}{}
-	for _, entry := range report.Stacks {
-		stackdir := entry.Stack.HostDir(cfg)
-		if stackdir == s.WorkingDir || strings.HasPrefix(stackdir, s.WorkingDir+string(filepath.Separator)) {
-			log.Debug().Msgf("selected stack: %s", entry.Stack.Dir)
+	filteredStacks, err := s.Engine.FilterStacks(report.Stacks, engine.ByWorkingDir(), engine.ByTags(s.Tags, s.NoTags))
+	if err != nil {
+		return err
+	}
 
-			selectedStacks[entry.Stack.Dir] = struct{}{}
-		}
+	selectedStacks := map[project.Path]struct{}{}
+	for _, entry := range filteredStacks {
+		selectedStacks[entry.Stack.Dir] = struct{}{}
 	}
 
 	vendorDir, err := s.Engine.VendorDir()
@@ -55,6 +55,7 @@ func (s *Spec) Exec(_ context.Context) error {
 		return err
 	}
 
+	cfg := s.Engine.Config()
 	results, err := generate.Load(cfg, vendorDir)
 	if err != nil {
 		return errors.E(err, "generate debug: loading generated code")